Extract code related to notification reasons in new module
