+++ apps/transport/lib/transport_web/templates/backoffice/page/form_dataset.html.heex
@@ -113,7 +113,7 @@
         </datalist>
         <%= label f, :reasons, class: "pt-12" do %>
           <%= dgettext("backoffice", "Notification reason") %>
-          <%= for reason <- DB.NotificationSubscription.subscribable_reasons_related_to_datasets(:producer) do %>
+          <%= for reason <- Transport.NotificationReason.subscribable_reasons_related_to_datasets(:producer) do %>
             <%= label f, reason do %>
               <%= checkbox(f, reason, hidden_input: false, value: true) %>
               <%= reason %>
